requirements.txt
opencv-python
scipy
Pillow
matplotlib
pyamg
scikit-image
numpy